Clean up namespace
