--- gbtis/AdminWindow.xaml.cs
@@ -1,93 +1,0 @@
-﻿using System;
-using System.Windows;
-using System.Windows.Media;
-
-namespace gbtis {
-    /// <summary>
-    /// Interaction logic for AdminWindow.xaml
-    /// </summary>
-    public partial class AdminWindow : Window {
-
-        // Window events
-        public event EventHandler Exit;
-        public event EventHandler Standby;
-        public event EventHandler Input;
-
-        /// <summary>
-        /// Initialize the window
-        /// </summary>
-        public AdminWindow() {
-            DataContext = this;
-            InitializeComponent();
-
-            Kinect kinect = Kinect.getInstance();
-            kinect.BitMapReady += BitMapArrived;
-        }
-
-        /// <summary>
-        /// Update the camera feed from the sensor
-        /// </summary>
-        /// <param name="img">The latest ImageSource</param>
-        void BitMapArrived(ImageSource img) {
-            this.Dispatcher.Invoke(new Action(() =>
-                sensorFeed.Source = img));
-        }
-
-        /// <summary>
-        /// Set the status bar's text
-        /// </summary>
-        /// <param name="text"></param>
-        public void SetStatus(String text) {
-            statusText.Text = text;
-        }
-
-        /// <summary>
-        /// Update the frame in the video feed
-        /// </summary>
-        /// <param name="frame">The new frame</param>
-        public void FrameArrived(ImageSource frame) {
-            sensorFeed.Source = frame;
-        }
-
-        /// <summary>
-        /// Close the program
-        /// </summary>
-        /// <param name="sender">Event source</param>
-        /// <param name="args">Event args</param>
-        private void FileExit_Click(object sender, EventArgs args) {
-            Exit?.Invoke(this, args);
-        }
-
-        /// <summary>
-        /// Display the about box
-        /// </summary>
-        /// <param name="sender">Event source</param>
-        /// <param name="args">Event args</param>
-        private void HelpAbout_Click(object sender, EventArgs args) {
-            MessageBox.Show(
-                Properties.Resources.aboutText,
-                Properties.Resources.aboutTitle,
-                MessageBoxButton.OK,
-                MessageBoxImage.Information
-            );
-        }
-
-        /// <summary>
-        /// Return to standby mode
-        /// </summary>
-        /// <param name="sender">Event source</param>
-        /// <param name="e">Event args</param>
-        private void FileStandby_Click(object sender, EventArgs args) {
-            Standby?.Invoke(this, args);
-        }
-
-        /// <summary>
-        /// Start the input mode
-        /// </summary>
-        /// <param name="sender">Event source</param>
-        /// <param name="e">Event args</param>
-        private void FileInput_Click(object sender, EventArgs args) {
-            Input?.Invoke(this, args);
-        }
-    }
-}
