--- gbtis/CanvasWindow.xaml
@@ -1,56 +1,0 @@
-﻿<Window x:Class="gbtis.CanvasWindow"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:gbtis"
-        mc:Ignorable="d"
-        
-        WindowStyle="None" ResizeMode="NoResize" Cursor="None"
-        WindowStartupLocation="CenterScreen" WindowState="Maximized">
-    <Grid>
-        <Grid.RowDefinitions>
-            <RowDefinition Height="*" />
-            <RowDefinition Height="8*" />
-            <RowDefinition Height="*" />
-        </Grid.RowDefinitions>
-
-        <Grid Grid.Row="0" Margin="10">
-            <Grid.ColumnDefinitions>
-                <ColumnDefinition Width="*" />
-                <ColumnDefinition Width="*" />
-                <ColumnDefinition Width="*" />
-                <ColumnDefinition Width="*" />
-            </Grid.ColumnDefinitions>
-
-           
-            <local:HoverButton x:Name="helpButton" x:FieldModifier="public" Grid.Column="0" Text="Help" Color="SlateGray" Margin="0,0,10,0" />
-            <local:HoverButton x:Name="clearButton" x:FieldModifier="public" Grid.Column="1" Text="Clear" Color="RoyalBlue" />
-            <local:HoverButton x:Name="cancelButton" x:FieldModifier="public" Grid.Column="2" Text="Cancel" Color="Crimson" />
-            <local:HoverButton x:Name="continueButton" x:FieldModifier="public" Grid.Column="3" Text="Continue" Color="ForestGreen" />
-           
-        </Grid>
-
-        <Border BorderBrush="Black" BorderThickness="0,1,0,1" Grid.Row="1">
-            <Grid>
-                <InkCanvas x:Name="canvas" EditingMode="Select" UseCustomCursor="True" Cursor="None">
-                    <InkPresenter x:Name="canvasPresenter" />
-                    <InkCanvas.DefaultDrawingAttributes>
-                        <DrawingAttributes Color="Black" FitToCurve="true" Width="10" Height="10"/>
-                    </InkCanvas.DefaultDrawingAttributes>
-                </InkCanvas>
-            </Grid>
-        </Border>
-
-        <Viewbox HorizontalAlignment="Stretch" Grid.Row="2">
-            <TextBlock Margin="10,0,10,0" x:Name="previewText" />
-        </Viewbox>
-
-        <Grid Grid.ColumnSpan="4" Grid.RowSpan="3">
-            <local:Cursor x:Name="cursor" HorizontalAlignment="Left" VerticalAlignment="Top"/>
-        </Grid>
-        <Image x:Name="sensorOverlay" Opacity="0.2" Margin="-0"
-               Stretch="UniformToFill" HorizontalAlignment="Center" VerticalAlignment="Center" 
-               Grid.ColumnSpan="3" Grid.RowSpan="10" />
-    </Grid>
-</Window>
