--- gbtis/Cursor.xaml
@@ -1,30 +1,0 @@
-﻿<UserControl x:Class="gbtis.Cursor"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
-             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:gbtis"
-             mc:Ignorable="d" 
-             d:DesignHeight="100" d:DesignWidth="100">
-    <Grid Width="100" Height="100">
-        <Border x:Name="erase" Background="White" BorderBrush="Black" BorderThickness="1">
-            <Grid x:Name="idle" HorizontalAlignment="Center" VerticalAlignment="Center">
-                <!-- Over canvas -->
-                <Line 
-                    X1="0" Y1="25" 
-                    X2="50" Y2="25" Stroke="Black" StrokeThickness="1" />
-                <Line 
-                    X1="25" Y1="0" 
-                    X2="25" Y2="50" Stroke="Black" StrokeThickness="1" />
-                
-                <!-- Over strokes -->
-                <Line 
-                    X1="0" Y1="25" 
-                    X2="50" Y2="25" Stroke="White" StrokeThickness="1" />
-                <Line 
-                    X1="25" Y1="0" 
-                    X2="25" Y2="50" Stroke="White" StrokeThickness="1" />
-            </Grid>
-        </Border>
-    </Grid>
-</UserControl>
