+++ gbtis/Windows/AdminWindow.xaml
@@ -1,4 +1,4 @@
-﻿<Window x:Class="gbtis.AdminWindow" x:Name="adminWindow"
+﻿<Window x:Class="gbtis.Windows.AdminWindow" x:Name="adminWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
