+++ gbtis/Windows/StandbyWindow.xaml
@@ -1,0 +1,53 @@
+﻿<Window x:Class="gbtis.Windows.StandbyWindow"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+        xmlns:local="clr-namespace:gbtis"
+        mc:Ignorable="d"
+        
+        Closing="Window_Closed"
+        WindowStyle="None" ResizeMode="NoResize"
+        WindowStartupLocation="CenterScreen" WindowState="Maximized">
+    <Grid>
+        <Grid.ColumnDefinitions>
+            <ColumnDefinition Width="*" />
+            <ColumnDefinition Width="*" />
+            <ColumnDefinition Width="*" />
+            <ColumnDefinition Width="*" />
+            <ColumnDefinition Width="*" />
+        </Grid.ColumnDefinitions>
+
+        <!-- Shows the current feed from the camera -->
+        <Border Background="DarkBlue" HorizontalAlignment="Stretch" VerticalAlignment="Stretch" Grid.Column="0" Grid.ColumnSpan="5">
+            <Image x:Name="sensorFeed" Stretch="UniformToFill" HorizontalAlignment="Center" VerticalAlignment="Center" />
+        </Border>
+
+        <!-- Shows a random selection of entries -->
+        <Border Background="White" Opacity="0.6" Grid.Column="0" Grid.ColumnSpan="2">
+            <Grid>
+                <Viewbox VerticalAlignment="Top" HorizontalAlignment="Stretch">
+                    <TextBlock x:Name="topName" Margin="5" Text="Loading..." HorizontalAlignment="Stretch" />
+                </Viewbox>
+                <Viewbox VerticalAlignment="Bottom" HorizontalAlignment="Stretch">
+                    <TextBlock x:Name="centerName" Margin="5" Text="Loading..." HorizontalAlignment="Stretch" TextAlignment="Right" />
+                </Viewbox>
+                <Viewbox VerticalAlignment="Center" HorizontalAlignment="Stretch">
+                    <TextBlock x:Name="bottomName" Margin="5" Text="Loading..." HorizontalAlignment="Stretch" TextAlignment="Right" />
+                </Viewbox>
+            </Grid>
+        </Border>
+
+        <!-- Intro text... Hidden until someone is in frame -->
+        <Viewbox HorizontalAlignment="Stretch" VerticalAlignment="Top" Grid.Column="2" Grid.ColumnSpan="3">
+            <Border>
+                <TextBlock x:Name="standbyMsg" Margin="10" FontWeight="Bold" Foreground="White" Text="Loading...">
+                    <TextBlock.Effect>
+                        <DropShadowEffect BlurRadius="2" Direction="0" ShadowDepth="0"/>
+                    </TextBlock.Effect>
+                </TextBlock>
+            </Border>
+        </Viewbox>
+
+    </Grid>
+</Window>
